Update the resampler to support native float input and use SIMD if possible. Extracting IResampler interface and making a ResamplerFactory for future potential native interop use. Fixing native lib package versions.
