+++ CSharp/Concentus.Native/Concentus.Native.nuspec
@@ -3,7 +3,7 @@
     <metadata>
 		<!-- Required elements-->
 		<id>Concentus.Native</id>
-		<version>1.5.1</version>
+		<version>1.5.2</version>
 		<description>Contains native Opus redistributable binaries for various platforms, intended as an optional addon to the core Concentus library. The default behavior for Concentus would be to look for these native binaries and use them if present, otherwise, fall back to a pure C# implementation of the codec. For best results, you should install this package ONLY into your top-level project, not into any intermediate libraries - this allows things like dotnet publish to correctly trim out native binaries that aren't relevant to your release target.</description>
 		<authors>Logan Stromberg</authors>
 		<icon>icon.png</icon>
