+++ CSharp/Concentus/Concentus.csproj
@@ -7,7 +7,7 @@
 	<SignAssembly>True</SignAssembly>
 	<AssemblyOriginatorKeyFile>..\Concentus.snk</AssemblyOriginatorKeyFile>
 	<GeneratePackageOnBuild>True</GeneratePackageOnBuild>
-	<Version>2.2.0</Version>
+	<Version>2.2.1</Version>
 	<Authors>Logan Stromberg</Authors>
 	<Description>This package is a portable C# implementation of the Opus audio compression codec (see https://opus-codec.org/ for more details). This package contains the Opus encoder, decoder, multistream codecs, repacketizer, as well as a port of the libspeexdsp resampler. It does NOT contain code to parse .ogg or .opus container files or to manage RTP packet streams. For better performance depending on your platform, see also the Concentus.Native package.</Description>
 	<Copyright>© Xiph.Org Foundation, Skype Limited, CSIRO, Microsoft Corp.</Copyright>
